added -f option
